+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
 - Make `/etc/fstab` be hardlinked inside the base environment.
+
+- Add support for specifying enabled systemd services inside the dpt file